+++ development/Sandbox/Providers/MicrosoftFoundry.cs
@@ -1,9 +1,24 @@
 using AgentFrameworkToolkit.MicrosoftFoundry;
 using AgentFrameworkToolkit.OpenAI;
+using Azure.AI.Extensions.OpenAI;
+using Azure.AI.Projects;
+using Azure.AI.Projects.Agents;
 using Azure.Identity;
+using Google.GenAI;
+using Google.Protobuf.WellKnownTypes;
 using Microsoft.Agents.AI;
+using Microsoft.Agents.AI.Foundry;
+using Microsoft.Extensions.Options;
 using OpenAI.Responses;
 using Secrets;
+using System.ClientModel;
+using AgentFrameworkToolkit;
+using AgentFrameworkToolkit.Tools.Common;
+using Microsoft.Extensions.AI;
+using Microsoft.Extensions.Logging;
+
+// ReSharper disable MethodHasAsyncOverload
+
 #pragma warning disable OPENAI001
 
 namespace Sandbox.Providers;
@@ -13,13 +28,26 @@ public static class MicrosoftFoundry
     public static async Task RunAsync()
     {
         Secrets.Secrets secrets = SecretsManager.GetSecrets();
-        MicrosoftFoundryConnection connection = new(secrets.MicrosoftFoundryEndpoint, new AzureCliCredential());
-
+        string endpoint = secrets.MicrosoftFoundryEndpoint;
+        MicrosoftFoundryConnection connection = new(endpoint, new AzureCliCredential());
         MicrosoftFoundryAgentFactory factory = new(connection);
-
-        //MicrosoftFoundryAgent agent2 = factory.DeclarativeAgentFactory.CreateAgent("myTest", "gpt-5.6-luna");
         /*
-        MicrosoftFoundryAgent agent3 = factory.DeclarativeAgentFactory.CreateAgent(new DeclarativeAgentOptions
+        MicrosoftFoundryAgent agent = factory.HostedAgentFactory.CreateAgent(new HostedAgentCreationOptions
+        {
+            DotNetRuntime = "dotnet_10",
+            Cpu = "0.5",
+            Memory = "1Gi",
+            SourceDirectory = @"C:\Users\rasmu\AppData\Local\Temp\AgentFrameworkToolkit\MyHostedAgent",
+            AssemblyName = "MyHostedAgent.dll",
+            Name = "Hosted",
+        });
+        */
+        MicrosoftFoundryAgent agent = factory.HostedAgentFactory.GetAgent("Hosted");
+
+        AgentResponse response = await agent.RunAsync("What is the Weather like in paris?");
+
+
+        MicrosoftFoundryAgent agent2 = factory.DeclarativeAgentFactory.CreateAgent(new DeclarativeAgentCreationOptions
         {
             Name = "MyCoolAgent",
             Model = "gpt-5.6-luna",
@@ -27,9 +55,17 @@ public static class MicrosoftFoundry
             ReasoningEffort = ResponseReasoningEffortLevel.Low,
             WebSearchTool = true,
             CodeInterpreterTool = true,
-            ReasoningSummaryVerbosity = ResponseReasoningSummaryVerbosity.Detailed
+            ReasoningSummaryVerbosity = ResponseReasoningSummaryVerbosity.Detailed,
+            Tools = [TimeTools.GetNowLocal()],
+
+            RawToolCallDetails = details =>
+            {
+                Console.WriteLine(details.ToString());
+            }
         });
-        */
+
+        AgentResponse agentResponse = await agent2.RunAsync("What is the time?");
+        Console.WriteLine(agentResponse);
         /*
         IList<MicrosoftFoundryAgent> agents = factory.DeclarativeAgentFactory.GetAgents();
 
@@ -43,7 +79,7 @@ public static class MicrosoftFoundry
         */
         //MicrosoftFoundryAgent agent = factory.DeclarativeAgentFactory.GetAgent("myTest", "1");
 
-        
+        /*
         MicrosoftFoundryAgent agent3 = factory.CreateAgent(new AgentOptions
         {
             Model = "gpt-5.6-luna",RawHttpCallDetails = details =>
@@ -52,20 +88,23 @@ public static class MicrosoftFoundry
                 Console.WriteLine(details.RequestData);
                 Console.WriteLine(details.ResponseData);
             }
-        });
+        });*/
         /*
         AIProjectClient client = connection.GetClient();
 
         ChatClientAgent agent = client.GetProjectOpenAIClient().GetProjectResponsesClient().AsAIAgent(model: "gpt-5.6-luna");
         */
-        AgentSession session = await agent3.CreateSessionAsync();
+
+
+
+        AgentSession session = await agent2.CreateSessionAsync();
 
         while (true)
         {
             Console.Write("> ");
             string input = Console.ReadLine() ?? "";
-            AgentResponse response = await agent3.RunAsync(input, session);
-            Console.WriteLine(response);
+            AgentResponse response2 = await agent2.RunAsync(input, session);
+            Console.WriteLine(response2);
 
             Console.WriteLine("------------------");
         }